chore: strip test code and move html-react-parser to dev dependency
