+++ packages/markdown/package.json
@@ -34,7 +34,6 @@
   "dependencies": {
     "@launchpad-ui/tokens": "workspace:~",
     "classix": "2.1.17",
-    "html-react-parser": "3.0.8",
     "isomorphic-dompurify": "^0.24.0",
     "marked": "^4.2.2"
   },
@@ -44,6 +43,7 @@
   },
   "devDependencies": {
     "@types/marked": "^4.0.3",
+    "html-react-parser": "3.0.8",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   }
